backend/requirements.txt
tensorflow==2.15.0
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.4
smtplib==0.0.1
email-validator==2.1.1